requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.0
defusedxml==0.7.1
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.49.0
fpdf2==2.7.8
idna==3.6
kombu==5.3.5
lxml==5.1.0
mysqlclient==2.2.1
numpy==1.24.4
pandas==2.0.3
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.13.0
